added colision detector (without pointy walls)

+++ src/Defines.h
@@ -2,8 +2,8 @@
 #define DEFINES_H
 #define ORIGIN { (float)texture.width / 2, (float)texture.height / 2 }
 
-constexpr float screenWidth = 800;
-constexpr float screenHeight = 800;
+constexpr int screenWidth = 800;
+constexpr int screenHeight = 800;
 const int tileSize = 22;
 
 #endif // DEFINES_H
